==> codes/R/06-plot_survival.R <==
# This code includes commands to generate figures of survival presented in the 
# manuscript.

# Initial setup ----
library(tidyverse)
library(ggfan)
library(RColorBrewer)
library(lubridate)
library(HDInterval)
library(ggdist)
library(gridExtra)
library(grid)

# Some of the data generated by the data processing file are needed
source("codes/R/02-process_fish_data.R")

out <- readRDS(paste("outputs/out_fit_dat.rds", sep = ""))
fit <- out$fit
jdat <- out$jdat
mcmc <- fit$samples
ns <- fit$mcmc.info$n.samples

# Figure 4 - Predictions of detection probability ----

## Make predictions ---

zrfl2 <- seq(min(jdat$zrfl2), max(jdat$zrfl2), 0.1)

det_prob_pred <- tibble(a_p = fit$sims.list$alpha.p,
                        b_zrfl2 = fit$sims.list$beta.p) %>%
  expand(nesting(a_p, b_zrfl2), zrfl2) %>% 
  mutate(p = plogis(a_p + b_zrfl2 * zrfl2),
         rfl2 = zrfl2 * sd(rf_dat$rainfall_l2_sdy) + mean(rf_dat$rainfall_l2_sdy))

pred_day <- select(det_prob_pred, a_p, b_zrfl2) %>%
  distinct() %>% 
  expand(nesting(a_p, b_zrfl2), 
         nesting(select(rf_dat, date, zrainfall_l2_sdy))) %>%
  mutate(p = plogis(a_p + b_zrfl2 * zrainfall_l2_sdy))

## Make plot of detection probability by rainfall ----

det_rfl2 <- ggplot(det_prob_pred, aes(x = rfl2, y = p)) +
   geom_fan(show.legend = FALSE) +
   ggfan::geom_interval(intervals = c(0, 0.95)) +
   annotate("text", x = 47.5, y = 0.45, label = "(a)", size = 5) +
   xlab("2-day lagged total daily rainfall (mm)") +
   ylab("NA") +
   scale_x_continuous(breaks = seq(0, 45, 5)) +
   scale_y_continuous(breaks = seq(0, 0.45, 0.05)) +
   scale_fill_gradientn(colours = rev(brewer.pal(n = 9, "Greys"))[-c(9)]) +
   theme_bw() +
   theme(legend.position = "none",
         panel.grid.major = element_line(colour = "gray95", linewidth = 0.25),
         panel.grid.minor = element_blank(),
         axis.text = element_text(size = 14),
         axis.title.x = element_text(size = 16, margin = margin(t = 12)),
         axis.title.y = element_text(size = 6, colour = "white"))

## Make plot of detection probability by date ----

det_date <- ggplot(pred_day, aes(x = date, y = p)) +
   geom_fan(show.legend = FALSE) +
   ggfan::geom_interval(intervals = c(0, 0.95), size = 0.25) +
   annotate("text", x = as.Date("2021-08-31"), y = 0.45, label = "(b)", size = 5) +
   xlab("Date") +
   ylab("NA") +
   scale_y_continuous(breaks = seq(0, 0.45, 0.05)) +
   scale_fill_gradientn(colours = rev(brewer.pal(n = 9, "Greys"))[-c(9)]) +
   theme_bw() +
   theme(legend.position = "none",
         panel.grid.major = element_line(colour = "gray95", linewidth = 0.25),
         panel.grid.minor = element_blank(),
         axis.text = element_text(size = 14),
         axis.title.x = element_text(size = 16, margin = margin(t = 12)),
         axis.title.y = element_text(size = 6, colour = "white"))

## Make plot of rainfall by date  ----

rfl_date <- ggplot(rf_dat, aes(x = date, y = rainfall_l0)) +
  geom_col() +
  annotate("text", x = as.Date("2021-08-31"), y = 50, label = "(C)", size = 5) +
  xlab("Date") +
  ylab("Total daily rainfall (mm)") +
  scale_y_continuous(expand = expansion(mult = c(0, 0.05))) +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.text = element_text(size = 14),
        axis.title.x = element_text(size = 16, margin = margin(t = 12)),
        axis.title.y = element_text(size = 16, margin = margin(r = 12)))

jpeg("plots/Figure_4.jpg", res = 300, unit = "cm", height = 30, width = 15)

grid.arrange(det_rfl2, det_date, rfl_date, nrow = 3)

dev.off()

# Figure 5 - Predictions of survival probability by site and date ----

## Make predictions ----

len_rgn <- select(res_dat, rel_dt, length) |>
  mutate(rel_dt = case_when(as_date(rel_dt) ==  ymd("2021-05-23") ~ "m23",
                            as_date(rel_dt) ==  ymd("2021-06-09") ~ "j09",
                            as_date(rel_dt) ==  ymd("2021-06-19") ~ "j19"),
         z_length = (length - mean(length)) / sd(length)) |>
  group_by(rel_dt) |>
  summarize(min = min(z_length),
            max = max(z_length))

nd_m23 <- expand.grid(rel_ste = c("lwr", "lke", "upr"),
                      rel_dt = "m23",
                      # Add zero to sequence to ensure we have predictions at 
                      # the mean
                      z_length = c(0, seq(filter(len_rgn, rel_dt == "m23")$min,
                                          filter(len_rgn, rel_dt == "m23")$max,
                                          0.1)))

nd_j09 <- expand.grid(rel_ste = c("lwr", "lke", "upr"),
                      rel_dt = "j09",
                      # Add zero to sequence to ensure we have predictions at 
                      # the mean
                      z_length = c(0, seq(filter(len_rgn, rel_dt == "j09")$min,
                                          filter(len_rgn, rel_dt == "j09")$max,
                                          0.1)))

nd_j19 <- expand.grid(rel_ste = c("lwr", "lke", "upr"),
                      rel_dt = "j19",
                      # Add zero to sequence to ensure we have predictions at 
                      # the mean
                      z_length = c(0, seq(filter(len_rgn, rel_dt == "j19")$min,
                                          filter(len_rgn, rel_dt == "j19")$max,
                                          0.1)))

nd <- rbind(nd_m23, nd_j09, nd_j19)

nd <- as.data.frame(model.matrix(~ z_length + rel_ste * rel_dt, nd)) %>%
  select(z_length:rel_dtj19) %>%
  rename(len = z_length, lke = rel_stelke, upr = rel_steupr,
         j09 = rel_dtj09, j19 = rel_dtj19) 

surv_pred <-  tibble(a_phi = fit$sims.list$alpha.phi,
                b_len = fit$sims.list$beta.phi[, 1],
                b_upr = fit$sims.list$beta.phi[, 2],
                b_lke = fit$sims.list$beta.phi[, 3],
                b_j09 = fit$sims.list$beta.phi[, 4],
                b_j19 = fit$sims.list$beta.phi[, 5],
                b_lke_j09 = fit$sims.list$beta.phi[, 6],
                b_lke_j19 = fit$sims.list$beta.phi[, 7],
                b_upr_j09 = fit$sims.list$beta.phi[, 8],
                b_upr_j19 = fit$sims.list$beta.phi[, 9]) %>%
  expand(nesting(a_phi, b_len, b_upr, b_lke, b_j09, b_j19, b_lke_j09,
                 b_lke_j19, b_upr_j09, b_upr_j19), 
         nesting(nd)) %>% 
  mutate(phi = plogis(a_phi + b_len  * len  +
                        b_lke * lke + b_upr * upr +
                        b_j09 * j09 + b_j19 * j19 + 
                        b_lke_j09 * lke * j09 + b_lke_j19 * lke * j19 +
                        b_upr_j09 * upr * j09 + b_upr_j19 * upr * j19),
         rel_ste = case_when(lke == 0 & upr == 0 ~ "Lower River",
                             lke == 1 ~ "Lake",
                             upr == 1 ~ "Upper River"),
         rel_ste = factor(rel_ste, levels = c("Lower River", "Lake", "Upper River")),
         rel_dt = case_when(j09 == 0 & j19 == 0 ~ "May 23",
                            j09 == 1 ~ "June 09",
                            j19 == 1 ~ "June 19"),
         rel_dt = factor(rel_dt, levels = c("May 23", "June 09", "June 19")))

# The surv_pred_summ data frame is used to populate the summaries presented in
# Table S1 of Supplement S2.
surv_pred_summ <- filter(surv_pred, len == 0) |>
  group_by(rel_ste, rel_dt) |>
  summarize(mdn = round(median(phi), 2),
            mad = round(mad(phi), 2),
            lwr = round(hdi(phi, credMass = 0.95)[1], 2),
            upr = round(hdi(phi, credMass = 0.95)[2], 2))

lbl <- data.frame(rel_dt = factor(c("May 23", "June 09", "June 19"),
                                  levels = c("May 23", "June 09", "June 19")),
                  label = c("(a)", "(b)", "(c)"))

jpeg("plots/Figure_5.jpg", res = 300, unit = "cm", height = 20, width = 30)

ggplot(filter(surv_pred, len == 0), aes(x = rel_ste, y = phi)) + 
  stat_gradientinterval(fill_type = "gradient", point_interval = "median_hdi",
                        fill = "gray40", .width = c(0.95), fatten_point = 2.5,
                        linewidth = 1.5) +
  geom_text(data = lbl, inherit.aes = FALSE, aes(x = 0.75, y = 1, label = label),
            size = 5) +
  scale_y_continuous(breaks = seq(0, 1, 0.1)) +
  ylab("Survival probability") +
  xlab("Release site") +
  facet_wrap(~ rel_dt) +
  theme_bw() +
  theme(legend.position = "none",
        legend.text = element_text(size = 12),
        panel.grid = element_blank(),
        axis.text.x = element_text(size = 12),
        axis.text.y = element_text(size = 12),
        axis.title.x = element_text(size = 16, margin = margin(t = 12)),
        axis.title.y = element_text(size = 16, margin = margin(r = 12)),
        strip.text = element_text(size = 14))

dev.off()


# Figure 6 - Predictions of survival by length ----

surv_pred_summ_len <- surv_pred %>% 
  group_by(rel_ste, rel_dt, len) %>%
  summarize(mdn = median(phi),
            lwr = hdi(phi, credMass = 0.95)[1],
            upr = hdi(phi, credMass = 0.95)[2]) %>%
  mutate(len_mm = len * sd(res_dat$length) + mean(res_dat$length))

jpeg("plots/Figure_6.jpg", res = 300, unit = "cm", height = 20, width = 30)

ggplot(surv_pred_summ_len, aes(x = len_mm, y = mdn, colour = rel_ste,
                               fill = rel_ste)) +
  geom_ribbon(aes(ymin = lwr, ymax = upr), alpha = 0.2, col = "white") +
  geom_line()  +
  geom_text(data = lbl, inherit.aes = FALSE, aes(x = 100, y = 1, label = label),
            size = 5) +
  xlab("Fork length (mm)") +
  ylab("Survival probability") + 
  labs(color = "Release Site", fill = "Release Site")+
  scale_y_continuous(breaks = seq(0, 1, 0.1),limits = c(0, 1))+
  scale_colour_manual(values = c("#DBA037", "#3A7BB3", "#CE7A3A")) +
  scale_fill_manual(values = c("#DBA037", "#3A7BB3", "#CE7A3A")) +
  facet_wrap(~ rel_dt) +
  theme_bw() +
  theme(legend.position = "top",
        legend.title = element_blank(),
        panel.grid = element_blank(),
        axis.title.y = element_text(size = 16, margin = unit(c(0, 5, 0, 0), "mm")),
        axis.title.x = element_text(size = 16, margin = unit(c(5, 0, 0, 0), "mm")),
        axis.text = element_text(size = 14),
        strip.text = element_text(size = 14),
        legend.text = element_text(size = 14))

dev.off()
