==> codes/R/08-plot_supplement_s2.R <==
# This code includes commands to generate figures of presented in Supplement S2

# Initial setup ----
library(tidyverse)
library(lubridate)
library(HDInterval)
library(ggdist)

# Some of the data generated by the data processing file are needed
source("codes/R/02-process_fish_data.R")

out <- readRDS(paste("outputs/out_fit_dat.rds", sep = ""))
fit <- out$fit
jdat <- out$jdat
mcmc <- fit$samples
ns <- fit$mcmc.info$n.samples

# Figure S1 - Parameters of freshwater residence model ----

## Coefficients ----

coef_res <- tibble(Intercept = fit$sims.list$alpha.r,
                   Length = fit$sims.list$beta.r[, 1],
                   `Upper River` = fit$sims.list$beta.r[, 2],
                   Lake = fit$sims.list$beta.r[, 3],
                   `June 09` = fit$sims.list$beta.r[, 4],
                   `June 19` = fit$sims.list$beta.r[, 5],
                   `Lake x June 09` = fit$sims.list$beta.r[, 6],
                   `Lake x June 19` = fit$sims.list$beta.r[, 7],
                   `Upper River x June 09` = fit$sims.list$beta.r[, 8],
                   `Upper River x June 19` = fit$sims.list$beta.r[, 9]) %>%
  pivot_longer(Intercept:`Upper River x June 19`, names_to = "parameter", 
               values_to = "posterior_sample") %>%
  # Use levels argument of factor() function to set the order in which to show
  # parameters on y-axis
  mutate(parameter = factor(parameter, levels = c("Intercept", "Length",
                                                  "Lake", "Upper River", "June 09",
                                                  "June 19", "Lake x June 09",
                                                  "Lake x June 19", 
                                                  "Upper River x June 09",
                                                  "Upper River x June 19"))) %>%
  group_by(parameter) %>%
  summarize(median = median(posterior_sample),
            lwr = hdi(posterior_sample, credMass = 0.95)[1],
            upr = hdi(posterior_sample, credMass = 0.95)[2])

plot_coef_res <- ggplot(coef_res, aes(x = median, y = parameter)) +
  geom_point(size = 2.5, colour = "black") +
  geom_errorbarh(aes(xmin = lwr, xmax = upr), height = 0, colour = "black") + 
  geom_vline(xintercept = 0, linetype = "dashed", colour = "gray50") +
  annotate("text", x = -3.45, y = 10.25, label = "(a)", size = 4) +
  xlab(NULL) +
  ylab(NULL) +
  scale_x_continuous(breaks = seq(-3.5, 3, 0.5)) +
  coord_cartesian(xlim = c(-3.35, 3)) +
  theme_bw() +
  theme(panel.grid = element_blank(),
        axis.title.x = element_text(size = 16, margin = unit(c(5, 0, 0, 0), "mm")),
        axis.text = element_text(size = 12))

## Shape parameters ----
shp_res <- tibble(`Lower River on May 23` = fit$sims.list$nu.r[, 1],
                  `Lake on May 23` = fit$sims.list$nu.r[, 2],
                  `Upper River on May 23` = fit$sims.list$nu.r[, 3],
                  `Lower River on June 09` = fit$sims.list$nu.r[, 4],
                  `Lake on June 09` = fit$sims.list$nu.r[, 5],
                  `Upper River on June 09` = fit$sims.list$nu.r[, 6],
                  `Lower River on June 19` = fit$sims.list$nu.r[, 7],
                  `Lake on June 19` = fit$sims.list$nu.r[, 8],
                  `Upper River on June 19` = fit$sims.list$nu.r[, 9]) %>%
  pivot_longer(`Lower River on May 23`:`Upper River on June 19`, names_to = "parameter", 
               values_to = "posterior_sample") %>% 
  # Use levels argument of factor() function to set the order in which to show
  # parameters on y-axis
  mutate(parameter = factor(parameter, levels = c("Lower River on May 23",
                                                  "Lake on May 23",
                                                  "Upper River on May 23",
                                                  "Lower River on June 09",
                                                  "Lake on June 09",
                                                  "Upper River on June 09",
                                                  "Lower River on June 19",
                                                  "Lake on June 19",
                                                  "Upper River on June 19"))) %>%
  group_by(parameter) %>%
  summarize(median = median(posterior_sample),
            lwr = hdi(posterior_sample, credMass = 0.95)[1],
            upr = hdi(posterior_sample, credMass = 0.95)[2])

plot_shp_res <- ggplot(shp_res, aes(x = median, y = parameter)) +
  geom_point(size = 2.5, colour = "black") +
  geom_errorbarh(aes(xmin = lwr, xmax = upr), height = 0, colour = "black") + 
  annotate("text", x = 0.4, y = 9.25, label = "(b)", size = 4) +
  xlab("Estimate") +
  ylab(NULL) +
  scale_x_continuous(breaks = seq(0.5, 7.5, 0.5)) +
  coord_cartesian(xlim = c(0.5, 7.5)) +
  theme_bw() +
  theme(panel.grid = element_blank(),
        axis.title.x = element_text(size = 16, margin = unit(c(5, 0, 0, 0), "mm")),
        axis.text = element_text(size = 12))


jpeg("plots/Figure_S2_S1.jpg", res = 300, unit = "cm", height = 20, width = 20)

grid.arrange(plot_coef_res, plot_shp_res, nrow = 2)

dev.off()

# Figure S2 - Parameters of survival model ----

coef_surv <- tibble(Intercept = fit$sims.list$alpha.phi,
                    Length = fit$sims.list$beta.phi[, 1],
                    `Upper River` = fit$sims.list$beta.phi[, 2],
                    Lake = fit$sims.list$beta.phi[, 3],
                    `June 09` = fit$sims.list$beta.phi[, 4],
                    `June 19` = fit$sims.list$beta.phi[, 5],
                    `Lake x June 09` = fit$sims.list$beta.phi[, 6],
                    `Lake x June 19` = fit$sims.list$beta.phi[, 7],
                    `Upper River x June 09` = fit$sims.list$beta.phi[, 8],
                    `Upper River x June 19` = fit$sims.list$beta.phi[, 9]) %>%
  pivot_longer(Intercept:`Upper River x June 19`, names_to = "parameter", 
               values_to = "posterior_sample") %>%
  # Use levels argument of factor() function to set the order in which to show
  # parameters on y-axis
  mutate(parameter = factor(parameter, levels = c("Intercept", "Length",
                                                  "Lake", "Upper River", "June 09",
                                                  "June 19", "Lake x June 09",
                                                  "Lake x June 19", 
                                                  "Upper River x June 09",
                                                  "Upper River x June 19"))) %>%
  group_by(parameter) %>%
  summarize(median = median(posterior_sample),
            lwr = hdi(posterior_sample, credMass = 0.95)[1],
            upr = hdi(posterior_sample, credMass = 0.95)[2])


jpeg("plots/Figure_S2_S2.jpg", res = 300, unit = "cm", height = 20, width = 30)

ggplot(coef_surv, aes(x = median, y = parameter)) +
  geom_point(size = 2.5, colour = "black") +
  geom_errorbarh(aes(xmin = lwr, xmax = upr), height = 0, colour = "black") + 
  geom_vline(xintercept = 0, linetype = "dashed", colour = "gray50") +
  xlab("Estimate") +
  ylab(NULL) +
  scale_x_continuous(breaks = seq(-2.5, 4.5, 0.5)) +
  theme_bw() +
  theme(panel.grid = element_blank(),
        axis.title.x = element_text(size = 16, margin = unit(c(5, 0, 0, 0), "mm")),
        axis.text = element_text(size = 12))

dev.off()

# Figure S3 - Size distribution of released and detected fish ----

# Make table of with sizes of fish released and detected
rel_size <- select(tag, rel_st, rel_dt, tag_id, length) |>
  mutate(rel_dt = case_when(day(rel_dt) == 23 ~ "May 23",
                            day(rel_dt) == 09 ~ "June 09",
                            day(rel_dt) == 19 ~ "June 19"),
         rel_dt = factor(rel_dt, levels = c("May 23", "June 09", "June 19")),
         type = "Released")

det_id <- det$tag_id

det_size <- filter(rel_size, tag_id %in% det_id) |>
  mutate(type = "Detected")

rel_det_size <- rbind(rel_size, det_size)

jpeg("plots/Figure_S2_S3.jpg", res = 300, unit = "cm", height = 20, width = 25)

ggplot(rel_det_size, aes(x = length, fill = type)) +
  geom_density(alpha = 0.4) +
  geom_text(data = lbl, inherit.aes = FALSE, aes(x = 102, y = 0.105, label = label),
            size = 5) +
  ylab("Density") +
  xlab(NULL) +
  facet_grid("rel_dt")+
  theme_bw()+
  coord_cartesian(xlim = c(60, 102)) +
  scale_color_manual(values = c("#DBA037",  "#3A7BB3")) +  
  scale_fill_manual(values = c("#DBA037",  "#3A7BB3")) +
  guides(fill = guide_legend(byrow = TRUE)) +
  theme(panel.grid = element_blank(),
        legend.position = c(0.085, 0.95),
        legend.title = element_blank(),
        legend.background = element_blank(),
        legend.text = element_text(size = 14),
        legend.spacing.y = unit(0.2, 'cm'),
        axis.title.y = element_text(size = 16, margin = unit(c(0, 5, 0, 0), "mm")),
        axis.title.x = element_text(size = 16, margin = unit(c(5, 0, 0, 0), "mm")),
        axis.text = element_text(size = 14),
        strip.text = element_text(size = 14))

dev.off()
